Making library more standards compliant
Moved the generateTranscript function out of the library, because it's a
user function and not part of the standard.

Renamed the constants in the TextTrack class so they can be used outside
of the class.

Added another html page to experiment with the IDL of TextTrack.

Added @default and @id attributes to the <track> element.